[migration-fixup] Rename build.deleted to build.untagged
