+++ alembic/versions/0337520adb1e_refactor_last_build_triggers.py
@@ -33,14 +33,14 @@ def upgrade():
             SELECT INTO pkg * FROM package WHERE id = pkg_id;
             SELECT INTO lb * FROM build
                 WHERE package_id = pkg_id
-                  AND NOT deleted
+                  AND NOT untagged
                 ORDER BY started DESC
                 LIMIT 1;
             IF lb.state = 2 THEN
                 SELECT INTO lcb * FROM build
                     WHERE package_id = pkg_id
                       AND (state = 3 OR state = 5)
-                      AND NOT deleted
+                      AND NOT untagged
                     ORDER BY started DESC
                     LIMIT 1;
             ELSE
@@ -92,7 +92,7 @@ def upgrade():
             EXECUTE PROCEDURE update_last_build_trigger();
         CREATE TRIGGER update_last_build_trigger_up
             AFTER UPDATE ON build FOR EACH ROW
-            WHEN (OLD.state != NEW.state OR OLD.deleted != NEW.deleted)
+            WHEN (OLD.state != NEW.state OR OLD.untagged != NEW.untagged)
             EXECUTE PROCEDURE update_last_build_trigger();
         CREATE TRIGGER update_last_build_trigger_del
             AFTER DELETE ON build FOR EACH ROW
